+++ client/front/lib/user_info.dart
@@ -7,7 +7,7 @@ import 'package:flutter_svg/svg.dart';
 import 'package:front/user_verify.dart';
 import 'package:http/http.dart' as http;
 import 'dart:convert';
-import 'sign_up_success.dart';
+import 'sign_up_complete.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
 
 //현재 화면에서 뒤로가기
